@@ -3,33 +3,22 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^6.2.1",
-		"@fortawesome/free-regular-svg-icons": "^6.2.1",
-		"@fortawesome/free-solid-svg-icons": "^6.2.1",
-		"@fortawesome/react-fontawesome": "^0.2.0",
-		"@testing-library/jest-dom": "^5.11.9",
-		"@testing-library/react": "^11.2.5",
-		"@testing-library/user-event": "^13.5.0",
-		"@typescript-eslint/parser": "^4.15.0",
-		"clsx": "^1.1.1",
-		"eslint-config-airbnb": "^18.2.1",
-		"eslint-plugin-jsx-a11y": "^6.4.1",
-		"history": "^5.2.0",
-		"howler": "^2.2.3",
-		"prop-types": "^15.7.2",
-		"react": "^17.0.1",
-		"react-avatar": "^3.10.0",
-		"react-dom": "^17.0.1",
-		"react-promise-tracker": "^2.1.0",
-		"react-router": "^6.2.1",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "^4.0.3",
-		"react-spinners": "^0.11.0",
-		"sinon-chai": "^3.7.0",
-		"styled-components": "^5.3.3",
-		"typescript": "^4.1.3",
-		"web-vitals": "^1.1.0",
-		"yup": "^0.32.11"
+		"@testing-library/jest-dom": "^5.16.5",
+		"@testing-library/react": "^14.0.0",
+		"@testing-library/user-event": "^14.4.3",
+		"@typescript-eslint/parser": "^5.59.7",
+		"eslint-config-airbnb": "^19.0.4",
+		"eslint-plugin-jsx-a11y": "^6.7.1",
+		"history": "^5.3.0",
+		"prop-types": "^15.8.1",
+		"react": "^18.2.0",
+		"react-avatar": "^5.0.3",
+		"react-dom": "^18.2.0",
+		"react-promise-tracker": "^2.1.1",
+		"react-router-dom": "^6.11.2",
+		"react-scripts": "^5.0.1",
+		"react-spinners": "^0.13.8",
+		"yup": "^1.2.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
